requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pdfplumber==0.11.4
google-generativeai==0.8.3
deep-translator==1.11.4
python-dotenv==1.0.1
html2image==2.0.5.2
Pillow==11.0.0
reportlab==4.2.5
